Move sharing urls into lib
