+++ packages/frontend/amp/components/topMeta/TopMetaPaidContent.tsx
@@ -10,7 +10,7 @@ import { PaidForBand } from '@frontend/amp/components/topMeta/PaidForBand';
 
 import { palette } from '@guardian/src-foundations';
 import { textSans, body } from '@guardian/pasteup/typography';
-import { getSharingUrls } from '@frontend/model/sharing-urls';
+import { getSharingUrls } from '@frontend/lib/sharing-urls';
 import { getAgeWarning } from '@frontend/model/age-warning';
 
 const headerStyle = css`
