+++ packages/frontend/web/components/ArticleHeader.tsx
@@ -8,7 +8,7 @@ import { getAgeWarning } from '@frontend/model/age-warning';
 import { pillarMap, pillarPalette } from '@frontend/lib/pillars';
 import { screenReaderOnly } from '@guardian/pasteup/mixins';
 import { headline, textSans, body } from '@guardian/pasteup/typography';
-import { getSharingUrls } from '@frontend/model/sharing-urls';
+import { getSharingUrls } from '@frontend/lib/sharing-urls';
 import {
     from,
     until,
